refactor: 完成领域模型重构与代码规范化
变更点：
- 新增实体基础类EntityBase统一实体公共属性
- 重构会话、标签、项目、任务等实体，引入NullableTime统一空时间处理
- 替换废弃的Session实体为UserSession并完善会话校验逻辑
- 拆分任务评论、检查项实体，替换原有的冗余实体定义
- 规范化领域服务、仓库接口命名，统一使用英文前缀命名
- 调整路由注释与代码结构，移除冗余注释
- 更新持久层模型与转换器适配新实体结构
- 重构分页、值对象代码，统一类型定义与校验逻辑
- 新增golangci-lint配置与vscode开发配置
- 修复用户角色默认值、状态类型等小bug

+++ domain/pomodoro/valueobjects/createPomodoro.go
@@ -5,6 +5,7 @@ import (
 	"time"
 )
 
+// CreatePomodoro 创建待办任务番茄工作请求
 type CreatePomodoro struct {
 	UserId      int64
 	SessionId   string
@@ -14,10 +15,11 @@ type CreatePomodoro struct {
 	Description string
 	StartAt     *time.Time
 	EndAt       *time.Time
-	Duration    int
+	Duration    uint8
 	Note        string
 }
 
+// Validate 验证创建待办任务番茄工作请求是否有效
 func (vo *CreatePomodoro) Validate() error {
 	if vo.SessionId == "" {
 		return errors.New("会话 ID 不能为空")
@@ -43,6 +45,7 @@ func (vo *CreatePomodoro) Validate() error {
 	return nil
 }
 
+// NewCreatePomodoro 创建待办任务番茄工作请求
 func NewCreatePomodoro(
 	userId int64,
 	sessionId string,
@@ -52,7 +55,7 @@ func NewCreatePomodoro(
 	description string,
 	startAt *time.Time,
 	endAt *time.Time,
-	duration int,
+	duration uint8,
 	note string,
 ) (*CreatePomodoro, error) {
 	vo := &CreatePomodoro{
